Changes to opcounters and namespacing
Updating opcounters to be roughly based on time. Higher order bits
of the opcounter are based on time since the epoch and lower order
bits are based on a rotating local counter. This breaks ties for
events that happen close together but provides a nice time-based
ordering for ids as long as no more than 1024 events occur at the
same time (we use Time.now.to_f, which gives sub-second resolution.)

Adding the prefix "flux:set" to all of our sorted sets so that
everything under the control of flux is consistently namespaced.