Don't halt event processing on an undefined attribute.
When an event matches a global handler (defined in config/schema.json), we
don't want to halt execution if one of the attributes in a single handler
is undefined. (If an attribute referenced in an inline handler is undefined,
we will still halt execution.)

As an example case, updated the following/unfollowing schema definitions to
accept either "followed" or "followee" as the followee parameter. We can
remove the former handler once client applications' code goes to production.