Explicitly requiring the yaml library when needed
We were pulling YAML in from somewhere, because this code used to
work. Somewhere between ruby-1.9.3-p194 and ruby-1.9.3-p286,
something changed that necessitates us explicitly requiring 'yaml'.

The default install for the chef cookbook I want to upgrade us to
is ruby-1.9.3-p327, which is why I came across this.